requirements.txt
setuptools
pyaudio
SpeechRecognition
--extra-index-url https://download.pytorch.org/whl/cu116
pytorch
numpy
git+https://github.com/openai/whisper.git
grpcio
